Split Environment into a separate module
